refactor(plugin-dev): move in-repo runJetWhale out of the published plugin
The published com.kitakkun.jetwhale plugin carried JetWhale-repo-internal build
logic: runJetWhale launched the local :jetwhale-host:app project via a resolvable
configuration. That reference has no meaning for external plugin authors and broke
their Gradle sync (the resolvable config pointed at a missing project).

Separate the concerns:
- Published plugin keeps only the external-facing tasks: packagePlugin,
  installPlugin, stageDevPlugin, downloadJetWhaleHost, runJetWhaleFromRelease
  (and the hostVersion extension). It no longer references :jetwhale-host:app, so
  external projects configure/sync cleanly by construction.
- New non-published gradle-conventions convention 'jetwhale-host-launch' adds
  runJetWhale (local :jetwhale-host:app launch), applied only by in-repo host
  modules (example/host). It reuses the published plugin's stageDevPlugin + dev dir.
- Drop hostApplicationProject from the extension (it was never overridden).

Verified: in-repo example:host exposes both runJetWhale and runJetWhaleFromRelease;
an external project (plugin from mavenLocal) exposes no runJetWhale and configures
with config cache, with no :jetwhale-host:app reference.
